Add deployment README.md
See 'README_Deploying_kb.pmem.io_to_GitHub_Pages_using_Travis-CI.md' for instructions on configuring Travis-CI and GitHub Pages to generate this site. We need to do this because this Jekyll site uses plugins not natively supported by GitHub Pages, so we build the site using Travis-CI and push the output to the 'gh-pages' branch on each push.